backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pymupdf>=1.24.0
openai>=1.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
